40 cleanup todos
* refactor some field names to end with 'Cost'

* do not allow deleting a player that is associated with a game

* resurrect seating notifications

* seating service use notification module instead of SMS connector

* resurrect sending emails

* update readm

+++ application/src/main/java/com/texastoc/module/player/service/PlayerService.java
@@ -4,19 +4,17 @@ import com.google.common.collect.ImmutableSet;
 import com.texastoc.common.AuthorizationHelper;
 import com.texastoc.exception.NotFoundException;
 import com.texastoc.exception.PermissionDeniedException;
+import com.texastoc.module.game.GameModule;
+import com.texastoc.module.game.GameModuleFactory;
+import com.texastoc.module.game.model.Game;
 import com.texastoc.module.notification.NotificationModule;
 import com.texastoc.module.notification.NotificationModuleFactory;
 import com.texastoc.module.player.PlayerModule;
+import com.texastoc.module.player.exception.CannotDeletePlayerException;
 import com.texastoc.module.player.exception.CannotRemoveRoleException;
 import com.texastoc.module.player.model.Player;
 import com.texastoc.module.player.model.Role;
 import com.texastoc.module.player.repository.PlayerRepository;
-import lombok.extern.slf4j.Slf4j;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
@@ -27,6 +25,11 @@ import java.util.Optional;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.RandomStringUtils;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Slf4j
 @Service
@@ -37,11 +40,13 @@ public class PlayerService implements PlayerModule {
   private final AuthorizationHelper authorizationHelper;
 
   private NotificationModule notificationModule;
+  private GameModule gameModule;
 
   // Only one server so cache the forgot password codes here
   private Map<String, String> forgotPasswordCodes = new HashMap<>();
 
-  public PlayerService(PlayerRepository playerRepository, BCryptPasswordEncoder bCryptPasswordEncoder, AuthorizationHelper authorizationHelper) {
+  public PlayerService(PlayerRepository playerRepository,
+      BCryptPasswordEncoder bCryptPasswordEncoder, AuthorizationHelper authorizationHelper) {
     this.playerRepository = playerRepository;
     this.bCryptPasswordEncoder = bCryptPasswordEncoder;
     this.authorizationHelper = authorizationHelper;
@@ -51,15 +56,15 @@ public class PlayerService implements PlayerModule {
   @Transactional
   public Player create(Player player) {
     Player playerToCreate = Player.builder()
-      .firstName(player.getFirstName())
-      .lastName(player.getLastName())
-      .email(player.getEmail())
-      .phone(player.getPhone())
-      .password(null)
-      .roles(ImmutableSet.of(Role.builder()
-        .type(Role.Type.USER)
-        .build()))
-      .build();
+        .firstName(player.getFirstName())
+        .lastName(player.getLastName())
+        .email(player.getEmail())
+        .phone(player.getPhone())
+        .password(null)
+        .roles(ImmutableSet.of(Role.builder()
+            .type(Role.Type.USER)
+            .build()))
+        .build();
 
     int id = playerRepository.save(playerToCreate).getId();
 
@@ -80,7 +85,8 @@ public class PlayerService implements PlayerModule {
   @Override
   @Transactional(readOnly = true)
   public List<Player> getAll() {
-    List<Player> players = StreamSupport.stream(playerRepository.findAll().spliterator(), false).collect(Collectors.toList());
+    List<Player> players = StreamSupport.stream(playerRepository.findAll().spliterator(), false)
+        .collect(Collectors.toList());
     Collections.sort(players);
     return players;
   }
@@ -112,16 +118,19 @@ public class PlayerService implements PlayerModule {
   @Transactional
   public void delete(int id) {
     verifyLoggedInUserIsAdmin();
-    // TODO call game service to see if player has any games
-//    if (player has any games) {
-//      throw new CannotDeletePlayerException("Player with ID " + id + " cannot be deleted");
-//    }
+
+    List<Game> games = getGameModule().getByPlayerId(id);
+    if (games.size() > 0) {
+      throw new CannotDeletePlayerException("Player with ID " + id + " cannot be deleted");
+    }
     playerRepository.deleteById(id);
   }
 
   @Override
   public void forgotPassword(String email) {
-    String generatedString = RandomStringUtils.random(5, 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z');
+    String generatedString = RandomStringUtils
+        .random(5, 'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M', 'N', 'O', 'P',
+            'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z');
     forgotPasswordCodes.put(email, generatedString);
     log.info("reset code: {}", generatedString);
     getNotificaionModule().sendEmail(Arrays.asList(email), "Reset Code", generatedString);
@@ -200,7 +209,8 @@ public class PlayerService implements PlayerModule {
   // verify the user is an admin
   private void verifyLoggedInUserIsAdmin() {
     if (!authorizationHelper.isLoggedInUserHaveRole(Role.Type.ADMIN)) {
-      throw new PermissionDeniedException("A player that is not an admin cannot update another player");
+      throw new PermissionDeniedException(
+          "A player that is not an admin cannot update another player");
     }
   }
 
@@ -214,7 +224,8 @@ public class PlayerService implements PlayerModule {
       }
       Player loggedInPlayer = players.get(0);
       if (loggedInPlayer.getId() != player.getId()) {
-        throw new PermissionDeniedException("A player that is not an admin cannot update another player");
+        throw new PermissionDeniedException(
+            "A player that is not an admin cannot update another player");
       }
     }
   }
@@ -224,5 +235,12 @@ public class PlayerService implements PlayerModule {
       notificationModule = NotificationModuleFactory.getNotificationModule();
     }
     return notificationModule;
+  }
+
+  private GameModule getGameModule() {
+    if (gameModule == null) {
+      gameModule = GameModuleFactory.getGameModule();
+    }
+    return gameModule;
   }
 }
